+++ isisd/isis_main.c
@@ -54,6 +54,7 @@
 #include "isisd/isis_zebra.h"
 #include "isisd/isis_te.h"
 #include "isisd/isis_errors.h"
+#include "isisd/isis_vty_common.h"
 
 /* Default configuration file name */
 #define ISISD_DEFAULT_CONFIG "isisd.conf"
@@ -207,6 +208,7 @@ int main(int argc, char **argv, char **envp)
 	prefix_list_init();
 	isis_init();
 	isis_circuit_init();
+	isis_vty_init();
 	isis_spf_cmds_init();
 	isis_redist_init();
 	isis_route_map_init();
